+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # PHPUNITKIT CHANGELOG
 
 * Fixed: Find next / previous failure keymaps (Windows)
+* Added [#55](https://github.com/gerardroche/sublime-phpunit/issues/55): Use different PHP executable
 * Added [#53](https://github.com/gerardroche/sublime-phpunit/issues/53): Vi keymaps are now disabled by default
 * Added [#52](https://github.com/gerardroche/sublime-phpunit/issues/52): Test File command
 * Changed: `phpunit.keymaps.vi` setting is now `phpunit.vi_keymaps` and can be enabled/disabled independently of the `phpunit.keymaps` setting.